chore: move crypto logic to pinia
